… .github/workflows/build-and-test.yml
… include/ttmlir/Target/TTNN/CMakeLists.txt
… lib/Dialect/TTNN/Transforms/CMakeLists.txt
… lib/Target/TTNN/TTNNToFlatbuffer.cpp
… runtime/include/tt/runtime/detail/test/ttnn/utils.h
… runtime/include/tt/runtime/detail/ttnn/ttnn.h
… runtime/include/tt/runtime/detail/ttnn/types.h
… runtime/include/tt/runtime/detail/ttnn/utils.h
… runtime/include/tt/runtime/runtime.h
… runtime/include/tt/runtime/tensor_cache.h
… runtime/include/tt/runtime/types.h
… runtime/include/tt/runtime/workarounds.h
… runtime/lib/binary.cpp
… runtime/lib/runtime.cpp
… runtime/lib/ttmetal/runtime.cpp
… runtime/lib/ttnn/operations/CMakeLists.txt
… runtime/lib/ttnn/operations/cache/load_cached.cpp
… runtime/lib/ttnn/operations/ccl/reduce_scatter.cpp
… runtime/lib/ttnn/operations/trace/trace.cpp
… runtime/lib/ttnn/operations/trace/trace.h
… runtime/lib/ttnn/operations/utils/utils.cpp
… runtime/lib/ttnn/program_executor.cpp
… runtime/lib/ttnn/runtime.cpp
… runtime/lib/ttnn/types/CMakeLists.txt
… runtime/lib/ttnn/types/layout_converter.cpp
… runtime/lib/ttnn/types/trace_cache.cpp
… runtime/lib/ttnn/types/types.cpp
… runtime/lib/ttnn/utils/utils.cpp
… runtime/python/binary/binary.cpp
… runtime/python/runtime/runtime.cpp