### Ticket
Closes #2703 

### Problem description
We want to support metal trace in our execution flow for perf improvements.

### What's changed
Added an e2e implementation of trace. Initially I wanted to target runtime only, but that turned out to be overly hacky, so I implemented a minimum update in the compiler to support this. I tried to make this as similar to const-eval as possible.

* Added `TTNN_TraceOp` that contains the information of the trace
* Added `TTNNTraceHoistTransform` pass that hoists ttnn operations into a separate function to trace. This pass is guarded by the command-line arg `enable-trace` in the ttnn pipeline
* Added lowering from TTNN to flatbuffer
* Added runtime op implementation, which involves a trace cache that is tied to the runtime device
  * This also helps ensure "thread safety" since executing two binaries on the same device in parallel is undefined behaviour anyway.

### Example

The following IR:
```
  func.func @matmul_with_bias(%arg0: tensor<64x32xbf16>, %arg1: tensor<32x64xbf16>, %arg2: tensor<64x64xbf16>) -> tensor<64x64xbf16> {
    // CHECK: ttnn.trace
    %0 = ttir.empty() : tensor<64x64xbf16>
    %1 = "ttir.matmul"(%arg0, %arg1, %0) : (tensor<64x32xbf16>, tensor<32x64xbf16>, tensor<64x64xbf16>) -> tensor<64x64xbf16>
    %2 = ttir.empty() : tensor<64x64xbf16>
    %3 = "ttir.add"(%1, %arg2, %2) : (tensor<64x64xbf16>, tensor<64x64xbf16>, tensor<64x64xbf16>) -> tensor<64x64xbf16>
    return %3 : tensor<64x64xbf16>
  }
```

Gets compiled into:

```
  func.func @matmul_with_bias_trace_0(%arg0: tensor<64x32xbf16, #ttnn_layout>, %arg1: tensor<32x64xbf16, #ttnn_layout1>, %arg2: tensor<64x64xbf16, #ttnn_layout2>) -> tensor<64x64xbf16, #ttnn_layout2> attributes {ttnn.trace} {
    %0 = "ttnn.matmul"(%arg0, %arg1) <{transpose_a = false, transpose_b = false}> : (tensor<64x32xbf16, #ttnn_layout>, tensor<32x64xbf16, #ttnn_layout1>) -> tensor<64x64xbf16, #ttnn_layout2>
    %1 = "ttnn.add"(%0, %arg2) : (tensor<64x64xbf16, #ttnn_layout2>, tensor<64x64xbf16, #ttnn_layout2>) -> tensor<64x64xbf16, #ttnn_layout2>
    "ttnn.deallocate"(%0) <{force = false}> : (tensor<64x64xbf16, #ttnn_layout2>) -> ()
    return %1 : tensor<64x64xbf16, #ttnn_layout2>
  }

  func.func @matmul_with_bias(%arg0: tensor<64x32xbf16, #ttnn_layout>, %arg1: tensor<32x64xbf16, #ttnn_layout1>, %arg2: tensor<64x64xbf16, #ttnn_layout2>) -> tensor<64x64xbf16, #ttnn_layout2> {
    %0 = "ttnn.get_device"() <{mesh_offset = #ttnn<mesh_offset 0x0>, mesh_shape = #ttnn<mesh_shape 1x1>}> : () -> !ttnn.device
    %1 = ttnn.trace(%0, 0, true, @matmul_with_bias_trace_0, [%arg0, %arg1, %arg2]) : (tensor<64x32xbf16, #ttnn_layout>, tensor<32x64xbf16, #ttnn_layout1>, tensor<64x64xbf16, #ttnn_layout2>) -> tensor<64x64xbf16, #ttnn_layout2>
    "ttnn.deallocate"(%arg2) <{force = false}> : (tensor<64x64xbf16, #ttnn_layout2>) -> ()
    "ttnn.deallocate"(%arg1) <{force = false}> : (tensor<32x64xbf16, #ttnn_layout1>) -> ()
    "ttnn.deallocate"(%arg0) <{force = false}> : (tensor<64x32xbf16, #ttnn_layout>) -> ()
    return %1 : tensor<64x64xbf16, #ttnn_layout2>
  }
```

This also works with const-eval:
```
func.func @matmul_with_bias_const_eval_0(%arg0: tensor<64x32xbf16, #ttnn_layout>, %arg1: tensor<32x64xbf16, #ttnn_layout1>) -> tensor<64x64xbf16, #ttnn_layout2> attributes {const_eval} {
  %0 = "ttnn.matmul"(%arg0, %arg1) <{transpose_a = false, transpose_b = false}> : (tensor<64x32xbf16, #ttnn_layout>, tensor<32x64xbf16, #ttnn_layout1>) -> tensor<64x64xbf16, #ttnn_layout2>
  return %0 : tensor<64x64xbf16, #ttnn_layout2>
}
func.func @matmul_with_bias_trace_0(%arg0: tensor<64x64xbf16, #ttnn_layout2>, %arg1: tensor<64x64xbf16, #ttnn_layout2>) -> tensor<64x64xbf16, #ttnn_layout2> attributes {ttnn.trace} {
  %0 = "ttnn.add"(%arg1, %arg0) : (tensor<64x64xbf16, #ttnn_layout2>, tensor<64x64xbf16, #ttnn_layout2>) -> tensor<64x64xbf16, #ttnn_layout2>
  return %0 : tensor<64x64xbf16, #ttnn_layout2>
}
func.func @matmul_with_bias(%arg0: tensor<64x32xbf16, #ttnn_layout> {tt.argument_type = #tt.argument_type<parameter>}, %arg1: tensor<32x64xbf16, #ttnn_layout1> {tt.argument_type = #tt.argument_type<parameter>}, %arg2: tensor<64x64xbf16, #ttnn_layout2> {tt.argument_type = #tt.argument_type<input>}) -> tensor<64x64xbf16, #ttnn_layout2> {
  %0 = "ttnn.get_device"() <{mesh_offset = #ttnn<mesh_offset 0x0>, mesh_shape = #ttnn<mesh_shape 1x1>}> : () -> !ttnn.device
  %1 = tt.load_cached(@matmul_with_bias_const_eval_0, [%arg0, %arg1]) : (tensor<64x32xbf16, #ttnn_layout>, tensor<32x64xbf16, #ttnn_layout1>) -> tensor<64x64xbf16, #ttnn_layout2>
  "ttnn.deallocate"(%arg1) <{force = false}> : (tensor<32x64xbf16, #ttnn_layout1>) -> ()
  "ttnn.deallocate"(%arg0) <{force = false}> : (tensor<64x32xbf16, #ttnn_layout>) -> ()
  %2 = ttnn.trace(%0, 0, true, @matmul_with_bias_trace_0, [%arg2, %1]) : (tensor<64x64xbf16, #ttnn_layout2>, tensor<64x64xbf16, #ttnn_layout2>) -> tensor<64x64xbf16, #ttnn_layout2>
  "ttnn.deallocate"(%1) <{force = false}> : (tensor<64x64xbf16, #ttnn_layout2>) -> ()
  "ttnn.deallocate"(%arg2) <{force = false}> : (tensor<64x64xbf16, #ttnn_layout2>) -> ()
  return %2 : tensor<64x64xbf16, #ttnn_layout2>
}
```

### Future improvements
The current implementation is more of a POC and by no means the final form of this feature. Some future tasks include:
* Add dynamic/multi-trace support, presumably integrating this into the optimizer
* Add support for multiple cqs, and l1 trace buffers
* Currently trace_region_size needs to be determined by the user in a trial-and-error fashion - keep increasing it until the trace fits. Could we potentially do this in op-model by repeatedly attempting to capture the trace until it fits on device, and bake the trace region size within the op?
* Add emitC support if needed. Currently the emitC pass asserts that `enable-trace` is not set
* Model tensor read-back in the compiler before executing trace if needed. Currently metal doesn't support device to device memcpy, and therefore trace needs to implicitly bounce the tensor off of host. I added a workaround flag in runtime to track this.
* Currently any callbacks that call ttnn ops that fall back to host will disrupt the trace and result in an error. This limits any golden callbacks from being hooked into runtime when trace is enabled. 

### Checklist
- [X] New/Existing tests provide coverage for changes
  - Added compiler/Dialect tests under `test/ttmlir/Dialect/TTNN/trace`
  - added `runtime/test/python/ttnn/device_agnostic/test_trace.py` that tests trace end to end with/without const eval