@svuckovicTT 2025-06-04
Took a high-level overview:
1. This would be really cool to discuss in a tech sync (maybe even outside of it if timing is not great).
2. I would _**really**_ like to have EmitC support from the get-go. It's fairly tiresome to always be playing catch-up. Think we can design this a little differently so that it works for all backends.