@odjuricicTT 2025-06-05
> I'd like to get this into tt-mlir first to verify functionality works correctly. Since this feature is disabled by default, it won't regress performance on the current main path.

I'm ok with not trying this out from FFE right away, but please file a followup issue for this as well. 

> This PR is already quite large, and benchmarking is a substantial task that would benefit from being documented as a separate effort. This would also let us properly track any optimizations or improvements needed based on the profiling results.

I was referring to just using existing e2e benchmarking infra to verify that the feature works as expected. You should be able to test this locally with ttrt:
```
ttrt run --benchmark --loop 32 flatbuffer_with_trace.tnn
```
This should print out the e2e time and save it to results.

Having this automated with some regression tests in CI is definitelly out of scope for this PR and is something we are working on.