@jnie-TT 2025-06-06
> 
> I was referring to just using existing e2e benchmarking infra to verify that the feature works as expected. You should be able to test this locally with ttrt

@odjuricicTT benchmarked through ttrt, and surprisingly seems like we are quite host bound on mnist linear:

### Without Trace
```
2025-06-06 19:43:18,006 - INFO - Total execution time over 32 loops: 111.22338601853698 ms
2025-06-06 19:43:18,006 - INFO - Execution time last loop: 3.447149007115513 ms
2025-06-06 19:43:18,006 - INFO - Batch size: 1
2025-06-06 19:43:18,006 - INFO - Samples per second: 290.0947994809121
```

### WIth Trace
```
2025-06-06 19:42:13,486 - INFO - Total execution time over 32 loops: 31.476844975259155 ms
2025-06-06 19:42:13,486 - INFO - Execution time last loop: 0.9687549900263548 ms
2025-06-06 19:42:13,486 - INFO - Batch size: 1
2025-06-06 19:42:13,486 - INFO - Samples per second: 1032.2527473874434
```

Seeing a 3.5x speedup when executing with trace