@odjuricicTT 2025-06-07
@jnie-TT Thanks! That's awesome!
Tho for bigger models like resnet the perf gain will probably not be so significant. The most important part for CNNs will be to add support for 2 command queues. I filed an issue for the followup here: https://github.com/tenstorrent/tt-mlir/issues/3698