@odjuricicTT 2025-06-07
@svuckovicTT It would be nice trace supported in emitC as well, tho i think that it makes sense to get it working e2e on the main path first.