@jnie-TT 2025-06-10
> So in the main emitc path (there's a test path as well, but it's equally easy to solve for), we have a pass that generates a driver function, which translates into main() function in C++. Today this calls the forward function of the model, among other things, and it would call the warmup as well in this scenario. The main function would then store the in/out tensor pointers.

@svuckovicTT I don't think I'm completely following this. How would the main function know which function to call (warm-up vs main loop)? Are you proposing both are called in the main function every time? That would re-capture the trace for all subsequent loops which is a waste of memory and performance. Also wouldn't the input output tensor pointers go out of scope if stored in the main function? I think we might be talking about two different execution flows. In my mind I'm imagining the user running the dylib API itself in a loop multiple times as we're doing in ttrt, instead of inserting the loop within the main function itself. If we were to insert the loop in the main function, does that mean the user would need to know how many times they want to execute during compile time and pass that into the compiler somehow?

> I'll have to disagree here, for several reasons:

If you look at the original [ticket](https://github.com/tenstorrent/tt-mlir/issues/2703#issuecomment-2884542338), I believe the consensus was to implement this as similar to consteval as possible. I'm open to re-discussing this if needed, but as you said this is just a POC, which in my understanding is just a small-scale, fundamental implementation that proves that this can work instead of a full fledged implementation that covers all aspects possible - I think it's reasonable to implement any further details on follow-ups. If we do decide to granularize the operations, that won't be a large change and will be seemless to users.

> Is there a reason to hurry this?

We should set up basic tests on the front-ends after we uplift this change and benchmark the performance differences to get a baseline understanding. The specific implementation details won't impact this initial testing. Once we start adding more complex functionality, having these fundamental tests in place will help us catch any regressions early. Also as this PR has been up for 2 weeks, I don't think we're necessarily hurrying this - I believe that for a POC it is in a mergeable state and we have good follow-ups to refine any details needed.